requirements.txt
Django==1.11.1
django-bootstrap-pagination==1.6.2
django-braces==1.11.0
django-constance==2.0.0
django-crequest==2016.3.16
django-crispy-forms==1.6.1
django-filter==1.0.4
django-picklefield==0.3.2
django-render-block==0.5
django-settings-export==1.2.1
django-tables2==1.6.1
django-templated-email==2.2.0
django-widget-tweaks==1.4.1
PyMySQL==0.7.11
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
